requirements.txt
kivy==2.1.0
requests==2.28.2
websocket-client==1.5.1
pillow==9.5.0
pyjnius==1.5.0  # Для Android
pyobjus==1.0.0  # Для iOS